move: dialect관련 패키지 ddl 외부로 이동 및 CreateTable DDL 변경
