+++ src/main/java/persistence/sql/ddl/builder/CreateQueryBuilder.java
@@ -1,8 +1,8 @@
-package persistence.sql.ddl.h2.builder;
+package persistence.sql.ddl.builder;
 
-import persistence.sql.ddl.QueryBuilder;
-import persistence.sql.ddl.h2.H2Dialect;
-import persistence.sql.ddl.h2.meta.TableName;
+import persistence.sql.ddl.h2.ColumnGenerator;
+import persistence.sql.dialect.H2Dialect;
+import persistence.sql.meta.TableName;
 
 public class CreateQueryBuilder implements QueryBuilder {
