+++ src/main/java/persistence/sql/ddl/builder/QueryBuilder.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.builder;
 
 public interface QueryBuilder {
     String generateSQL(final Class<?> clazz);
