+++ src/main/java/persistence/sql/dialect/column/Nullable.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.h2.meta;
+package persistence.sql.dialect.column;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Id;
