+++ src/test/java/persistence/sql/ddl/h2/builder/CreateQueryBuilderTest.java
@@ -3,7 +3,8 @@ package persistence.sql.ddl.h2.builder;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import persistence.domain.Person;
-import persistence.sql.ddl.QueryBuilder;
+import persistence.sql.ddl.builder.QueryBuilder;
+import persistence.sql.ddl.builder.CreateQueryBuilder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -19,11 +20,10 @@ class CreateQueryBuilderTest {
         assertThat(query).isEqualTo(
                 "create table users\n" +
                         "(\n" +
-                        "    id bigint NOT NULL generated by default as identity,\n" +
+                        "    id bigint NOT NULL generated by default as identity PRIMARY KEY,\n" +
                         "    nick_name varchar(255) NULL,\n" +
                         "    old int NULL,\n" +
-                        "    email varchar(255) NOT NULL,\n" +
-                        "    primary key (id)\n" +
+                        "    email varchar(255) NOT NULL\n" +
                         ");\n");
     }
 }
