+++ src/test/java/persistence/sql/ddl/h2/meta/ColumnNameTest.java
@@ -7,6 +7,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import persistence.sql.meta.ColumnName;
 
 import java.lang.reflect.Field;
