+++ src/test/java/persistence/sql/ddl/h2/meta/H2ColumnTest.java
@@ -19,7 +19,7 @@ class H2ColumnTest {
     void generatePK() throws NoSuchFieldException {
         H2Column emailH2Column = new H2Column(personClass.getDeclaredField("id"));
 
-        assertThat(emailH2Column.generateColumnSQL()).isEqualTo("id bigint NOT NULL generated by default as identity");
+        assertThat(emailH2Column.generateColumnSQL()).isEqualTo("id bigint NOT NULL generated by default as identity PRIMARY KEY");
     }
 
     @Test
